requirements.txt
psutil==5.8.0
mss==6.1.0
tqdm==4.64.1
Pymem==1.8.5
python-magic==0.4.24